Fixed the folding mechanism issue
when calling the async(places(0), ()->fold(fold) in run(), there was an ambiguity issue between the local fold and the distant fold. It was resolved by the compiler by using the distant fold which is not what we wanted. Problem fixed by renaming the local variable to circumvate the issue